Challenge-100Days/Day1-Function-Call/requirements.txt
openai==1.106.1
python-dotenv==1.1.1
